% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress.R
\name{lz4_compress}
\alias{lz4_compress}
\title{Compress a raw vector}
\usage{
lz4_compress(src)
}
\arguments{
\item{src}{raw vector containing data to be compressed.}
}
\value{
raw vector of compressed data
}
\description{
Compress a raw vector
}
