% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lz4.R
\name{lz4_decompress}
\alias{lz4_decompress}
\title{Decompess a raw vector into the original data.}
\usage{
lz4_decompress(src)
}
\arguments{
\item{src}{raw vector of compressed data}
}
\value{
uncompressed vector
}
\description{
The \code{src} vector must be a raw vector created by \code{lz4_compress} as
it relies on the bespoke framing created there.  This function will not
decompress an object compressed with another LZ4 implemetation e.g. \code{fst}
}
