% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress.R, R/lz4.R
\name{lz4_uncompress}
\alias{lz4_uncompress}
\title{Uncompress a raw vector into the original data.}
\usage{
lz4_uncompress(src)

lz4_uncompress(src)
}
\arguments{
\item{src}{raw vector of compressed data}
}
\value{
uncompressed vector

uncompressed vector
}
\description{
The \code{src} vector must be a raw vector created by \code{lz4_compress} as
it relies on the bespoke framing created there.  This function will not
decompress an object compressed with another LZ4 implementation e.g. \code{fst}

The \code{src} vector must be a raw vector created by \code{lz4_compress} as
it relies on the bespoke framing created there.  This function will not
decompress an object compressed with another LZ4 implementation e.g. \code{fst}
}
