% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress.R, R/lz4.R
\name{lz4_compress}
\alias{lz4_compress}
\alias{lz4hc_compress}
\title{Compress a vector of raw, integer, real or logical values.}
\usage{
lz4_compress(src, acceleration = 1)

lz4hc_compress(src, level = 9L)

lz4_compress(src, acceleration = 1)

lz4hc_compress(src, level = 9L)
}
\arguments{
\item{src}{source vector containing data to be compressed.
Data must be raw, integer, real or logical only.}

\item{acceleration}{Acceleration, Default 1.  Higher values run faster but with
lower compression. This value is ignored if \code{use_hc = TRUE}}

\item{level}{compression level for LZ4HC compression. default 9L. Valid range 3-12.}
}
\value{
raw vector of compressed data

raw vector of compressed data
}
\description{
Compress a vector of raw, integer, real or logical values.

Compress a vector of raw, integer, real or logical values.
}
