% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress.R
\name{lz4_decompress}
\alias{lz4_decompress}
\title{Decompress a raw vector}
\usage{
lz4_decompress(src)
}
\arguments{
\item{src}{raw vector of compressed data}
}
\value{
uncompressed vector
}
\description{
The \code{src} vector must be a raw vector created by \code{lz4_compress}
}
