% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{lz4_serialize_stream}
\alias{lz4_serialize_stream}
\alias{lz4_unserialize_stream}
\title{Serialize to file, raw or connection}
\usage{
lz4_serialize_stream(x, dst, acc = 1L)

lz4_unserialize_stream(src)
}
\arguments{
\item{x}{R object}

\item{dst}{filename (character), raw()/NULL for raw vector}

\item{acc}{accerlation. Default 1. Valid range [1, 65535].  Higher values
means faster compression, but larger compressed size.}

\item{src}{data source for unserialization. file, or raw_vector()}
}
\description{
Serialize to file, raw or connection
}
